Friday_Bot/requirements.txt
appdirs==1.4.3
astroid==2.3.3
certifi==2020.4.5.1
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
filetype==1.0.5
future==0.18.2
imageio==2.1.2
imageio-ffmpeg==0.4.1
importlib-metadata==1.6.0
importlib-resources==1.4.0
imutils==0.5.3
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
InstagramAPI==1.0.2
instapy-cli==0.0.12
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
moviepy==0.2.3.2
networkx==2.4
numpy==1.18.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.1.1
proglog==0.1.9
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.6.1
PyWavelets==1.1.1
pyzbar==0.1.8
requests==2.11.1
requests-toolbelt==0.7.0
scikit-image==0.16.2
scipy==1.4.1
selenium==3.141.0
six==1.14.0
tornado==6.0.4
tqdm==4.11.2
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.17
wget==3.2
wrapt==1.11.2
zipp==3.1.0
